requirements.txt
       fastapi
       gunicorn
       uvicorn
       jinja2
       pytest
       requests
       pre-commit
       python-dotenv
       python-multipart
       aiofiles
       pillow
